refactor: Separate integrations by type again
